script: Rewrite the manager stuff
The GameState class was added, which represent the one global singleton based on the
ScriptableObject. All other managers are also a ScriptableObject based (but not
ScriptableObjectSingleton based). Therefor some classes were updated to relfect those changes.
The one major change is that GameManager was split and new BootsrapManager was created which is an
initialization only manager.