script: Create EditorAssetManager
EditorAssetManager is used solely in the editor context. It was created in the Ja2Core.asmdef,
because it could be used in some runtime code run during editor play mode, therefor, putting it to
the Ja2Editor.asmdef wouldn't be an option and the code would need to be duplicated in the
AssetManager.